Much better solution to inventory variant inventory quantity problem - #343. Adds tests, and fixes issue where parent/grandparent quantities were not adding up properly. (11 Squashed) Squashed commits: [272ddf1] Make publications more consistent. Fixes #395. [be04b1c] Allow database ids as valid slug for tags.
Bevor this change it was for example not possible to use a product id
as name (slug) for a tag, because the application interpreted the id as
a tag id, instead of a tag name.

A valid use case would be to display a link to a product's related products.
Allowing a product's id as tag name (slug) it is easy to build an URL to related products e.g.
product/tag/<product_id> .

The old behaviour is still in service as a fallback.
[9cf1232] remove testing logging.
[97a4a5f] Update inventory calculator again. Passes all tests against inventory calculation, is less brittle, and faster.
[207c0ad] WIP: Still need some work to update all variants if/when the inventory starts wrong.
[c360ae1] Fix issue where inventory was undefined if variant was new option
[2beb2a0] Compensate for bootstrap's padding when applying .container-fluid class.
[0055b68] Replace brittle inventory update hook with slightly more robust update hook that will update all ancestors in parent chain.
[7f6a91c] Update inventoryVariant tests to test creation of variable quantities of inventoryVariants being added.
[192b7c5] Addl Inline doc for cloneVariant method.