Subscription callback was not executed in a predictable way.
Use subsciption.ready() instead.

Without subscription.ready() the settings.public part of the social
package was not assigned reliably to _self.socialSettings, because the
callback was not always executed.

The result was that the social icons did not appear in the footer.

With subscription.ready() _self.socialSettings is assigned in a reliable
way which in turn then
triggers template rendering to display the social icons in the footer.