Apply Meteor method call from GraphQL resolver
Adds a new `applyWithContext` function that wraps meteor calls with the
GraphQL context. The context includes the Meteor user. The function sets
the Meteor context for the invocation. In the future the context could
also hold other info, like the connection.

We could use `Meteor.call` and splat the ...args, but `apply` is used to
for one-to-one semantic match and to clarify that we are passing args as
an array.