chore: update .snyk file to ignore cryptiles
There's a new vulnerability reported in the outdated @google-cloud/storage@1.1.1
which is introduced by a dependency on reaction-file-collections' dependency tus-node-server.
I'm adding an ignore as there's a PR that has been accepted to resolve this dependency vulnerability.
Additionally, Reaction Commerce does not execute the code that has the vulnerability in this case.